Hw5

+++ scrapper/src/main/java/edu/java/dto/response/LinkResponse.java
@@ -4,4 +4,4 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import java.net.URI;
 
 @Schema
-public record LinkResponse(int id, URI url) {}
+public record LinkResponse(long id, URI url) {}
